Fix rollup creation in `Storage#perf_capture`
Because of the change in Rails 4 to `Model#attributes` which returns a
hash keyed by strings instead of symbols, a bug was introduced in
`#perf_capture` when merging attributes. It can be illustrated by the
following example:

```ruby
class Person < ActiveRecord::Base
  # persons table has id, name
end

attrs = {:name => "Alice"}
attrs.reverse_merge!(Person.new.attributes)
attrs # => {:name => "Alice", "name" => nil}
Person.new(attrs).name # => nil
```

The fix here is to call `#symbolize_keys` on `Person#attributes` so that
the `reverse_merge` won't merge in a nil value.

To better isolate and test where this was going wrong a utility method
was extracted that moves the crucial hash manipulation code into it. It
is otherwise identical to the code in `#perf_capture` apart from the
added `symbolize_keys` and can be seen failing without it.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1267390