Add deprecated warnings when passing parse_ws_string string arguments
parse_ws_string always accepted arguments as a hash
And in the case of this bug: sending multiple '|' delimited
arguments as a hash solves the original problem.  This fix
adds deprecation warnings to parse_ws_string when arguments
are sent as a string - noting the solution in the warning text.