Bearer type authentication doesn't need userid
Bearer type authentication should not need a userid to be present. The
code that assigns the authentication to the provider made certain
assumptions about the userid - that its presence indicated a change, and
its absence indicated a delete. This updates it to ensure that an
authentication can be added without a userid.

The change also removes the password from the spec's openshift
credentials, because this was never needed (the UI just provides fields
for username and token).

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1284005