core: fix start_event_monitor logic
In some cases `@tid` could be dead (not `alive?`) but still have a
`status`, which would lead to an infinite loop.

In fact:

    @tid ||= start_event_monitor

would never `start` a new `event_monitor` (since `@tid` already has
value), and:

    if @tid.status.nil?
      dead_tid, @tid = @tid, nil
      ...

would never take place as `@tid.status` is not `nil` resulting in an
infinite loop.

Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>