Using chrony for time sync instead of ntpdate
ntpdate is to be retired and chronyd is the new default
time sync daemon on Centos 7 and has been running on the
appliance along side our code running ntpdate. This opens
up the possibility of trying to sync with both the timeservers
configured in `/etc/chrony.conf` which have always been set to the
defaults and `/etc/ntp.conf` which is what is edited when the user
changes ntp servers.

Also the motivation behind using a custom script which ran ntpdate
was to have the schedule worker protected from large jumps in time
by dropping an exit file and waiting for it to exit.

The schedule worker has not responded to the exit file since
commit 85d34a9a869ce2eefcf69d7c58644a24cc2359ba was merged in 2012.

https://bugzilla.redhat.com/show_bug.cgi?id=1275677