Fix memoized helper race condition
Upon upgrading to RSpec 3, this spec completely hung forever. After spelunking
in the automation engine and RSpec, I've determined that the mocking here seems
to be caused by a thread deadlock since the thread safety additions made here:

https://github.com/rspec/rspec-core/commit/ffe00a1d4e369e312881e6b2c091c8b6fb7e6087
(note: This is now a part of rspec-support)

I'm not entirely understanding specifics of the issue beyond knowing that the
lazy-evaluation of using { } attempts to lock the mutex which is already locked
and held by the main thread, and thus two threads are now waiting for each
other to be evaluated.

Today I learned that using .and_return instead of a block to receive() is
actually eager loaded, and thus allows us to sidestep this issue (or we could
just use a variable, of course)