Remove region scoping from models excluded from replication
The original intention of scoping models to a region was so that
the global region didn't access rows that didn't belong to it.
This mattered specifically for seeded tables where the global
region would have its own versions of the same rows.

This idea was conceived before the ability to exclude tables from
replication was added. Because we can now exclude the tables that we
don't want the global region to see there is no longer any reason to
set a default scope on these tables as no database will ever have
rows from more than one region.

Original commit sha 0ac04a1d033f0aeeea45ad5344b1752726aad300