Return empty command lines for permission issues or already exited pids
We should consistently return strings so the best we can do is return an empty
string for these unhappy paths.

We can't take action for empty command lines, so in the case of exited pids, this
should be ok.  Permission issues are only a problem on development environments
but not for appliances.

```
[NoMethodError]: undefined method `include?' for nil:NilClass  Method:[rescue in monitor]
../gems/pending/util/miq-process.rb:163:in `is_worker?'
```