Remove redefinition of FORMAT_SUB_TYPES constant
This is causing an error in seeds as MiqExpression is being redefined in
one rake task and needed as normal in another. If you remove the
redefinition, loading ui_constants.rb fails as it depends on the loading
of MiqExpression's FORMAT_SUB_TYPES constant, which requires loading...(etc etc).

Constants should arguably _not_ be used this way, but that's out of
scope for this fix. The number of occurrences of FORMAT_SUB_TYPES in the
frontend does not merit the small shortcut redefining the constant as a
helper provides, so for now just prefix where needed and remove the
shortcut.

This change:
* Removes the redefinition of FORMAT_SUB_TYPES from ui_constants.
* Prefixes FORMAT_SUB_TYPES for the front end properly.
* Removes the redefinition of MiqExpression (fixes rake tasks/seeds).