Fix that state of host filter is not visbile in UI
Fixing bad matching of objects in belongs to filters in belongsto2object_list.
After this belongsto2object_list returns Host object moreover, so
we need to skip it here for origin behavior.

This settings is in group, in Host/Nodes& Clusters tab.
When host was checked it was stored in db but not visible in
tree that it is checked.