Upgrade RubyZip Gem to 1.2
The rubyzip gem has been locked at 0.9.5 for 4 years due to a namespace conflict
caused by a change in 0.9.7.  The api changed in 1.0 and there are other gems that
rely on the new API, including winrm-fs, which is needed for winrm-elevated.

Adding gem ZipZip allows older api code to continue to function alongside code using
the newer api.

Spec tests against the new gem which failed with 0.9.7 pass.