Remove Metric::Common#apply_time_profile_daily
In the past, daily records could be queried on the fly via adhoc rollups
in VimPerformanceDaily, and so this method was needed to filter those
records on the Ruby side.  Once we removed the adhoc daily rollups, and
relied solely on a TimeProfile to get daily data, this code stopped
being used.

The code for apply_time_profile is still needed for hourly records
however, since those can still be queried on the fly.