Remove unused refresh_state since refreshes use the ems. :fire:
This was added back in early 2008 in acf394cfc704 when we must have sent the
request through the smart proxy residing on the host.  I think ???

Unless my git fu is off, this method became unused in b78174781baa back in
February 2008, when we moved to using the ems to refresh the states.