Add migration to add id column and primary key to join tables
This migration also creates the sequence using the current region.

Because of this, records in a global (replicated) environment need
to be removed as we will not be able to guarentee that the new
primary key assigned in the regional database is set to the same
row in the global database at the time of the migration.

Also, only delete rows on the global region if they came from a different region

This will preserve rows that belong to the global region
like ones in seeded tables like `miq_groups_users` or if
someone is managing some environment using the global database.