Add ability to update a subscription using #save or #save!
In order to allow updates where the password is not passed
in from the UI the password is pulled from the dsn stored for the
subscription record when the record is being updated.

A nil in the password field indicates that the password was not
changed, but an empty string will attempt to set the password to
an empty string.