Reset WinRM connection when WMI fails due to Max Operations Limit
The WinRM gem currently assumes that a Windows server allows 1500
operations per connection, and resets the connection just before hitting
the limit.

Unfortunately, this limit is configurable and a WMI error may be returned
by any operation.  Issue /https://github.com/WinRb/WinRM/issues/197
tracks this problem.  The gem owner will eventually fix this issue by watching
for the error and simply resetting the connection when it occurs, but he
won't be doing so until version 2 of the gem - which is a while off.

We are propagating the fix up a level to the MiqWinRM wrapper class here.