Add a script to remove out of date rubyrep triggers
When a table which is being replicated is renamed the rubyrep
trigger should be dropped and recreated to ensure only rows
referencing the new table are inserted into rr_pending_changes.

If this is not done properly, triggers referencing the old table
name will remain functional on the renamed table. This has been
seen to cause the replicate process to time out when it happens with
high-churn tables.

https://bugzilla.redhat.com/show_bug.cgi?id=1323951