Callback before automation starts
If some of the provision objects implement a before_ae_starts
automate engine will call it before it starts the Automate Engine.
This is the bookend for the 'after_ae_delivery' which is run after
automation ends. It passes in the options hash that it was called with.
The object getting the options can analyze for retries.