Production migration fail due to cached schema for model
Production migration fails due to cached schema for model. Model
contains old schema, so the field parent_ems_id, that we created
couple of migrations back, will not be in the model
ExtManagementSystem and it will fail on create!. Seems like
this appears only on production env, where the schema is not
reloaded. Also this occurs only when we do e.g. upgrade, so
when migration has some data, before we continue with other
migrations. This migration is data migration touching OpenStack.

Fixes BZ:
https://bugzilla.redhat.com/show_bug.cgi?id=1329137