Handle gracefully when cloud -> network relation gets deleted
Handle gracefully when CloudManager -> NetworkManager relation
gets deleted. This should not happened since both creation
and deletion of cloud_manager is always packed with network
manager in a transaction. And yet, I had reported this occurred.

Which would probably point to some lowlevel consistency bug.
It has not been reproduced yet.

This PR allows us to display and delete disconnected CloudManager
and NetworkManager, without presenting any 500 error to the
user.