Add db connection checkin/checkout debug logging.
Related to #7801

Now that we use puma, which handles requests in threads, and we have multiple
services making requests to puma, the ops UI, self service UI, and the REST API,
we need logging of connection pool usage to try to track down misconfiguration or
a bug.

For example, if we have need to configure the puma min/max thread count (config/puma.rb)
(5,5 currently) to work correctly with the rails connection pool size of 5,
we need a way to know if we have too many/few puma threads or connection pool size.

```
[----] D, [2016-04-29T12:26:55.359811 #31844:3fc79545e204] DEBUG -- : PostgreSQLAdapter#log_after_checkout, connection_pool: size: 5, connections: 1, in use: 1, waiting_in_queue: 0
[----] D, [2016-04-29T12:26:59.263333 #31844:3fc79545e204] DEBUG -- : PostgreSQLAdapter#log_after_checkin, connection_pool: size: 5, connections: 1, in use: 0, waiting_in_queue: 0
```