Make root_tenant scope to the current region
Previously in a global database seeding would attach global region
groups to tenants from replicated regions.

When the replicated data was removed due to uninstalling replication
on the region (rubyrep) or removing the subscription (pglogical)
the link from the group to tenant would be broken and many things
would fail until the database was reseeded.