evm.pid / process existence is the safe way to ask if we're running?
When we changed the process title using setproctitle to "MIQ Server" [1], we broke
rake evm:start's detection of a previously running server.

While a process title/command line is somewhat helpful to check, only the server
should be writing to evm.pid so checking for a process from evm.pid is enough.

https://bugzilla.redhat.com/show_bug.cgi?id=1335391

[1] 1ba7801ee1d2591f17fe05861d726ac0d9a20fa1